% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_topo}
\alias{county_topo}
\title{TopoJSON file displaying georeferenced boundaries of U.S. Counties.}
\format{A TopoJSON characer vector}
\source{
\url{}
}
\usage{
county_topo
}
\description{
Description
}
\keyword{datasets}

