% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_recolor.R
\name{topo_recolor}
\alias{topo_recolor}
\title{change color scales by 'count' value in topoJson}
\usage{
topo_recolor(topo, max, FUN)
}
\arguments{
\item{topo}{topoJSON character vector with \code{count} attribute}

\item{max}{maximum value of \code{count} attribute}

\item{function}{defining color mapping such as \code{leaflet::colorNumeric}}
}
\value{
topoJSON character vector
}
\description{
change color scales by 'count' value in topoJson
}
\examples{
topo_recolor(topo_prac, 300, palfx)
}

